+++ src/Backend/Program.cs
@@ -9,6 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Imlight.Common.Utilities;
+using Imlight.Server.Data;
 using Imlight.Server.Login;
 using Imlight.Server.Database;
 using Imlight.Server.Patch;
