+++ src/Server/Data/CoreObjectFactory.cs
@@ -6,20 +6,16 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Timers;
-using static WizUnraveler.Cache.TypeCache;
-using WizUnraveler;
-using Imlight.Common.Utilities;
 using Imlight.Common.Cryptography;
+using Imlight.Common.Utilities;
+using Imlight.Server.Database;
 using Imlight.Server.Shared.Secrets;
-using WizUnraveler.Cache;
 using WizUnraveler.Common;
 using WizUnraveler.IO;
-using WizUnraveler.ObjectProperty;
+using static WizUnraveler.Cache.TypeCache;
 
-namespace Imlight.Server.Database
+namespace Imlight.Server.Data
 {
     public static class CoreObjectFactory
     {
