+++ src/Server/Data/Records/Character/EquipmentSlot.cs
