+++ src/Server/Game/GameWorld.cs
@@ -6,6 +6,7 @@
 using System.Collections.Generic;
 using Akka.Actor;
 using Imlight.Common.Utilities;
+using Imlight.Server.Data;
 using Imlight.Server.Database;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
