+++ src/Server/Game/Models/Character.cs
@@ -1,0 +1,6 @@
+namespace Imlight.Server.Game.Models;
+
+public class Character
+{
+    
+}
