+++ src/Server/Game/Zone/WizardZoneLoader.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using Akka.Actor;
 using Imlight.Common.Utilities;
+using Imlight.Server.Data;
 using Imlight.Server.Database;
 using Imlight.Server.Shared.Packets;
 using SharpDX;
